Move GoogleOfferingParser to common module for tests
