+++ purchases/src/main/kotlin/com/revenuecat/purchases/OfferingParserFactory.kt
@@ -3,7 +3,7 @@ package com.revenuecat.purchases
 import androidx.annotation.VisibleForTesting
 import com.revenuecat.purchases.common.OfferingParser
 import com.revenuecat.purchases.common.errorLog
-import com.revenuecat.purchases.google.GoogleOfferingParser
+import com.revenuecat.purchases.common.GoogleOfferingParser
 
 @VisibleForTesting(otherwise = VisibleForTesting.PACKAGE_PRIVATE)
 object OfferingParserFactory {
